* 'develop' of github.com:RocketChat/Rocket.Chat: (63 commits)
  fix: Messages jumping after reaction (#28770)
  fix: Quotes chain off by one error in quote chain limit settings (#28281)
  docs: update documentation (#28719)
  chore: /help notifying the shortcuts in multiple messages (#28765)
  refactor: remove Users from fibers 14 (#28780)
  regression(file-upload:s3): Fix missing store import (#28781)
  refactor: remove Users from fibers 13 (#28772)
  refactor: remove Users from fibers 12 (#28767)
  ci: revert Variable Setup changes (#28754)
  feat: auto link custom domain (#28501)
  refactor: `Rooms` model 4/4 (#28742)
  regression: invalid callback callAsync (#28766)
  refactor: remove Users from fibers 11 (#28740)
  refactor: Convert our Upload Code to TS (#28711)
  refactor: Move callbacks to ts 1X (#28710)
  refactor: Remove Users from fibers 10 (#28737)
  fix: message-parser links (#28735)
  refactor(client): Miscelaneous (#28732)
  refactor: `Rooms` model 3/4 (#28736)
  refactor: `Rooms` model 2X (#28715)
  ...

+++ apps/meteor/app/lib/server/functions/setUserAvatar.ts
@@ -14,23 +14,23 @@ export function setUserAvatar(
 	contentType: string,
 	service: 'rest',
 	etag?: string,
-): void;
+): Promise<void>;
 export function setUserAvatar(
 	user: Pick<IUser, '_id' | 'username'>,
 	dataURI: string,
 	contentType?: string,
 	service?: 'initials' | 'url' | 'rest' | string,
 	etag?: string,
-): void;
-export function setUserAvatar(
+): Promise<void>;
+export async function setUserAvatar(
 	user: Pick<IUser, '_id' | 'username'>,
 	dataURI: string | Buffer,
 	contentType: string | undefined,
 	service?: 'initials' | 'url' | 'rest' | string,
 	etag?: string,
-): void {
+): Promise<void> {
 	if (service === 'initials') {
-		Users.setAvatarData(user._id, service, null);
+		await Users.setAvatarData(user._id, service, null);
 		return;
 	}
 
@@ -104,7 +104,7 @@ export function setUserAvatar(
 	);
 
 	const fileStore = FileUpload.getStore('Avatars');
-	fileStore.deleteByName(user.username);
+	user.username && fileStore.deleteByName(user.username);
 
 	const file = {
 		userId: user._id,
@@ -114,10 +114,10 @@ export function setUserAvatar(
 
 	const result = fileStore.insertSync(file, buffer);
 
-	const avatarETag = etag || result?.etag || null;
+	const avatarETag = etag || result?.etag || '';
 
-	Meteor.setTimeout(function () {
-		Users.setAvatarData(user._id, service, avatarETag);
+	Meteor.setTimeout(async function () {
+		await Users.setAvatarData(user._id, service, avatarETag);
 		void api.broadcast('user.avatarUpdate', {
 			username: user.username,
 			avatarETag,
